+++ hmily-tac/hmily-tac-sqlparser/hmily-tac-sqlparser-model/src/main/java/org/dromara/hmily/tac/sqlparser/model/statement/dml/HmilyDeleteStatement.java
@@ -19,7 +19,6 @@ package org.dromara.hmily.tac.sqlparser.model.statement.dml;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.dromara.hmily.tac.sqlparser.model.segment.dml.assignment.SetAssignmentSegment;
 import org.dromara.hmily.tac.sqlparser.model.segment.dml.predicate.WhereSegment;
 import org.dromara.hmily.tac.sqlparser.model.segment.generic.table.SimpleTableSegment;
 
@@ -28,21 +27,19 @@ import java.util.LinkedList;
 import java.util.Optional;
 
 /**
- * Update statement.
+ * Delete statement.
  */
 @Getter
 @Setter
-public final class UpdateStatement extends HmilyDMLStatement {
+public final class HmilyDeleteStatement extends HmilyDMLStatement {
     
     private final Collection<SimpleTableSegment> tables = new LinkedList<>();
-    
-    private SetAssignmentSegment setAssignment;
     
     private WhereSegment where;
     
     /**
      * Get where.
-     * 
+     *
      * @return where segment
      */
     public Optional<WhereSegment> getWhere() {
